refactor: CardAliasPage 커스텀훅으로 정리
