+++ src/hooks/useCardAliasPage.ts
@@ -1,0 +1,45 @@
+import { ChangeEvent, useContext, useEffect, useRef } from 'react';
+
+import CardInfoContext from '../contexts/CardInfoContext';
+
+import { PATHNAME } from '../constants/pathname';
+import { useNavigationTo } from './useNavigationTo';
+
+export const useCardAliasPage = () => {
+  const {
+    cardNumbers,
+    expirationDate,
+    ownerName,
+    securityCode,
+    password,
+    cardCompany,
+    cardAlias,
+    setCardAlias,
+    setCardInfoList,
+  } = useContext(CardInfoContext);
+  const buttonRef = useRef<HTMLButtonElement>(null);
+  const { handleClick } = useNavigationTo(PATHNAME.HOME);
+
+  useEffect(() => {
+    if (cardAlias && cardAlias?.length >= 15) {
+      buttonRef.current?.focus();
+    }
+  }, [cardAlias]);
+
+  const onChange = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
+    setCardAlias(value);
+  };
+
+  const onClick = () => {
+    setCardInfoList(prev => {
+      return [
+        ...prev,
+        { cardNumbers, expirationDate, ownerName, securityCode, password, cardCompany, cardAlias },
+      ];
+    });
+
+    handleClick();
+  };
+
+  return { onChange, onClick, cardAlias, buttonRef };
+};
